requirements.txt
streamlit
scenedetect
opencv-python-headless
google-generativeai
pandas
